Authorization ownership and logger

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task-status.enum';
@@ -15,8 +16,8 @@ export class TasksService {
   ) {}
 
   /* Get task by ID */
-  async getTaskById(id: number): Promise<TaskEntity> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<TaskEntity> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!found) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
@@ -24,20 +25,20 @@ export class TasksService {
   }
 
   /* Delete task by ID */
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" does not exist`);
     }
   }
 
   /* Get all task */
-  async getTasks(filterDto: GetTasksFilterDto): Promise<TaskEntity[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<TaskEntity[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async updateTaskEstatusById(id: number, status: TaskStatus): Promise<TaskEntity> {
-    const task = await this.getTaskById(id);
+  async updateTaskEstatusById(id: number, status: TaskStatus, user: User): Promise<TaskEntity> {
+    const task = await this.getTaskById(id, user);
     /* just update the task object */
     task.status = status;
     await task.save();
@@ -45,7 +46,7 @@ export class TasksService {
   }
 
   //Create new task
-  async createTask(createTaskDto: CreateTaskDto): Promise<TaskEntity> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<TaskEntity> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 }
